reed/catalog/serializers.py
from rest_framework import serializers

from .models import Book


class BookSerializer(serializers.ModelSerializer):
    class Meta:
        model = Book
        fields = ['isbn13', 'title', 'description', 'publisher', 'authors']
